fixing changes

+++ src/components/groups/group.js
@@ -5,30 +5,45 @@ import { Link } from 'react-router-dom'
 
 import Auth from '../../lib/auth'
 import CreateCategory from '../categories1/createCategory.js'
+import EditGroups from './editGroups'
 
 class Group extends React.Component {
   constructor() {
     super()
 
     this.state = {
-      groups: []
+      createdBy: {}
     }
     this.handleDelete = this.handleDelete.bind(this)
+    this.reRender = this.reRender.bind(this)
+
   }
 
   handleDelete() {
-    axios.delete(`/api/groups/${this.props._id}`,
-      { headers: { Authorization: `Bearer ${Auth.getToken()}`}})
-      .then(() => this.props.history.push('/groups'))
-      .catch(err => console.log(err.message))
+    if (this.props.createdBy === Auth.getPayload().sub) {
+      axios.delete(`/api/groups/${this.props._id}`,
+        { headers: { Authorization: `Bearer ${Auth.getToken()}`}})
+        .then(() => this.props.rerender())
+        .catch(err => console.log(err.message))
+    } else {
+      console.log('not yours!')
+    }
+  }
+  reRender() {
+    const { edit } = this.state
+    this.setState({ edit: !edit })
   }
 
-  isOwner() {
-    return Auth.isAuthenticated() && this.props.createdBy === Auth.getPayload().sub
+  removeFromState(){
+
   }
 
   render() {
-    console.log(`Created by = ${this.props.createdBy}, Payload = ${Auth.getPayload().sub},  Authorised = ${Auth.isAuthenticated()},  Group name = ${this.props.name} `)
+    console.log(this.props)
+    if (this.state.edit) {
+      return  <EditGroups  group={this.props}/>
+    }
+    console.log(this.state.createdBy, 'createdBy')
     return(
       <div className="card">
         <div className="card-header">
@@ -39,12 +54,19 @@ class Group extends React.Component {
         <div className="card-content">
           <CreateCategory groupId={this.props._id}>Create Category</CreateCategory>
         </div>
+        <div className="card-footer">
+        </div>
+
         <footer className="card-footer">
-            {this.isOwner() && <button className="card-footer-item is-onethird" onClick={this.reRender}>Edit</button>}
-            {this.isOwner() && <button className="card-footer-item is-onethird" onClick={this.handleDelete}>Delete</button>}
-          <Link to={`/groups/${this.props._id}`} className="card-footer-item is-onethird">
+
+          <button className="card-footer-item is-onethird" onClick={this.reRender}>Edit</button>
+
+          <button className="card-footer-item is-onethird" onClick={this.handleDelete}>Delete</button>
+
+          <Link to={`/groups/${this.props._id}/tasks`} className="card-footer-item is-onethird">
             <strong>Tasks</strong>
           </Link>
+
         </footer>
       </div>
     )
